+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Seam carving".
 - Tests for "Rational number tree".
 - Tests for "The optimal urinal problem".
+- Tests for "Light bulbs".
 
 ## [1.11.0] - 2022-07-31
 ### Added